fix(reuse): restore storage state
